chore: Update GitHub repository URL in base.html
This commit updates the GitHub repository URL in the base.html file to reflect the new repository name "TotalityCRM" instead of "ClientesCrudFlask". The URL is now pointing to "https://github.com/MarioCarvalhoBr/TotalityCRM". This change ensures that the link to the repository is accurate and up to date.

Note: Please remove any meta information such as issue references, tags, or author names before using the commit message.